[fix](fe)change session variable group_concat_max_len from int to long (#15515)

+++ fe/fe-core/src/main/java/org/apache/doris/qe/SessionVariable.java
@@ -656,7 +656,7 @@ public class SessionVariable implements Serializable, Writable {
     public int repeatMaxNum = 10000;
 
     @VariableMgr.VarAttr(name = GROUP_CONCAT_MAX_LEN)
-    public int groupConcatMaxLen = 2147483646;
+    public long groupConcatMaxLen = 2147483646;
 
     // If this fe is in fuzzy mode, then will use initFuzzyModeVariables to generate some variables,
     // not the default value set in the code.
